requirements.txt
# AI Stock Advisor — Python Dependencies
# Install with: pip install -r requirements.txt

# Data & Math
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Deep Learning
tensorflow>=2.13.0          # or tensorflow-cpu if no GPU

# Financial Data
yfinance>=0.2.36

# Visualisation
plotly>=5.17.0

# AI Chatbot
google-generativeai>=0.7.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0        # for loading .env files with API keys
